Move the StateMachineId to a different file
